file extension takes precedence over file-type
When a file is executed in Windows, Windows looks for the file extension
rather than the file type, and thus we should show similar behavior.

An example has been shown by Oliver Matula and Christopher Scheuring in
their recent Hack in the Box talk [1] using a technique well-documented
by Ange Albertini.

As a free bonus this also makes it easier in the future to add new file
extensions and as usual makes the code cleaner.

[1]: http://haxpo.nl/materials/haxpo2015ams/D3%20-%20Oliver%20Matula%20and%20Christopher%20Scheuring%20-%20Evaluating%20the%20APT%20App%20Armor.pdf